README.md
This repository is the final landing page created in the course for Front-End Web UI Frameworks and Tools: Bootstrap 4 on Coursera. Feel free to know how the final project will be looking if you have just started this course. There are overall 4 easy assignments. The code in this repository is for the final project after the assignment 4.